[breadboard-web/-ui] Move -ui wrapper to -web
