+++ packages/breadboard-web/public/styles/global.css
@@ -51,6 +51,7 @@ html {
   --bb-icon-error: url(/third_party/icons/error.svg);
   --bb-icon-expand: url(/third_party/icons/expand.svg);
   --bb-icon-fit: url(/third_party/icons/fit.svg);
+  --bb-icon-frame-reload: url(/third_party/icons/frame-reload.svg);
   --bb-icon-history: url(/third_party/icons/history.svg);
   --bb-icon-info: url(/third_party/icons/info.svg);
   --bb-icon-input: url(/third_party/icons/input.svg);
