+++ packages/breadboard-web/src/preview.ts
@@ -1,0 +1,7 @@
+/**
+ * @license
+ * Copyright 2023 Google LLC
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+console.log("Preview");
